services/accounts/pyproject.toml
[project]
name = "accounts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asyncpg==0.31.0",
    "certifi==2026.1.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography>=46.0.5",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.128.0",
    "fastapi-cli==0.0.20",
    "fastapi-cloud-cli==0.11.0",
    "fastar==0.8.0",
    "greenlet==3.3.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "psycopg==3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.0",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt>=2.11.0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.22",
    "pyyaml==6.0.3",
    "rich==14.3.1",
    "rich-toolkit==0.17.1",
    "rignore==0.7.6",
    "sentry-sdk==2.50.0",
    "shellingham==1.5.4",
    "sqlalchemy==2.0.46",
    "starlette==0.50.0",
    "typer==0.21.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uvicorn==0.40.0",
    "watchfiles==1.1.1",
    "websockets==16.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest-asyncio>=1.3.0",
]
